Renaming visualizations plugin folder to kebab-case
