+++ src/plugins/visualizations/public/expression-functions/vis_dimension.ts
