+++ src/plugins/visualizations/public/saved-visualizations/find_list_items.test.ts
@@ -33,7 +33,7 @@
 import { findListItems } from './find_list_items';
 import { coreMock } from '../../../../core/public/mocks';
 import { SavedObjectsClientContract } from '../../../../core/public';
-import { VisTypeAlias } from '../vis_types';
+import { VisTypeAlias } from '../vis-types';
 
 describe('saved_visualizations', () => {
   function testProps() {
