+++ src/plugins/visualizations/public/saved-visualizations/find_list_items.ts
@@ -33,8 +33,8 @@
 import _ from 'lodash';
 import { SavedObjectAttributes, SavedObjectsClientContract } from '../../../../core/public';
 import { SavedObjectLoader } from '../../../../plugins/saved_objects/public';
-import { VisTypeAlias } from '../vis_types';
-import { VisualizationsAppExtension } from '../vis_types/vis_type_alias_registry';
+import { VisTypeAlias } from '../vis-types';
+import { VisualizationsAppExtension } from '../vis-types/vis_type_alias_registry';
 
 /**
  * Search for visualizations and convert them into a list display-friendly format.
