+++ src/plugins/visualizations/public/saved-visualizations/saved_visualization_references.ts
