+++ src/plugins/visualizations/public/saved-visualizations/saved_visualizations.ts
@@ -35,7 +35,7 @@ import {
 } from '../../../../plugins/saved_objects/public';
 import { findListItems } from './find_list_items';
 import { createSavedVisClass } from './_saved_vis';
-import { TypesStart } from '../vis_types';
+import { TypesStart } from '../vis-types';
 
 export interface SavedObjectOpenSearchDashboardsServicesWithVisualizations
   extends SavedObjectOpenSearchDashboardsServices {
