+++ src/plugins/visualizations/public/vis-types/base_vis_type.ts
