+++ src/plugins/visualizations/public/vis-types/vis_type_alias_registry.ts
