+++ src/plugins/visualizations/public/vis.test.ts
@@ -46,7 +46,7 @@ jest.mock('./services', () => {
   }
 
   // eslint-disable-next-line
-  const { BaseVisType } = require('./vis_types/base_vis_type');
+  const { BaseVisType } = require('./vis-types/base_vis_type');
   // eslint-disable-next-line
   const { SearchSource } = require('../../data/common/search/search_source');
   // eslint-disable-next-line
