+++ src/plugins/visualizations/public/wizard/search-selection/search_selection.tsx
@@ -37,7 +37,7 @@ import React from 'react';
 import { IUiSettingsClient, SavedObjectsStart } from '../../../../../core/public';
 
 import { SavedObjectFinderUi } from '../../../../../plugins/saved_objects/public';
-import { VisType } from '../../vis_types';
+import { VisType } from '../../vis-types';
 
 interface SearchSelectionProps {
   onSearchSelected: (searchId: string, searchType: string) => void;
