+++ src/plugins/visualizations/public/wizard/type-selection/new_vis_help.test.tsx
