+++ src/plugins/visualizations/public/wizard/type-selection/vis_type_icon.tsx
